With -delete, don't complain about non-empty directories.  Otherwise "cd /tmp; find . -mtime +7 -delete" is excessively noisy.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@18720 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ usr.bin/find/function.c
@@ -972,7 +972,7 @@ f_delete(plan, entry)
 
 	/* rmdir directories, unlink everything else */
 	if (S_ISDIR(entry->fts_statp->st_mode)) {
-		if (rmdir(entry->fts_accpath) < 0)
+		if (rmdir(entry->fts_accpath) < 0 && errno != ENOTEMPTY)
 			warn("-delete: rmdir(%s)", entry->fts_path);
 	} else {
 		if (unlink(entry->fts_accpath) < 0)
